requirements.txt
openai-whisper
transformers
torch
pydub
ffmpeg-python
scikit-learn
